move integration tests to example
