+++ example/models/__init__.py
@@ -1,0 +1,7 @@
+from example.models.system import System
+from example.models.telemetry import Telemetry
+
+__all__ = [
+    "System",
+    "Telemetry"
+]
