+++ example/models/system.py
@@ -1,0 +1,26 @@
+from typing import Any, Dict
+from myosin import StateModel
+
+
+class System(StateModel):
+
+    def __init__(self, id) -> None:
+        super().__init__(id)
+
+    @property
+    def online(self) -> bool:
+        return self.__online
+
+    @online.setter
+    def online(self, status: bool) -> None:
+        self.__online = status
+
+    def serialize(self) -> Dict[str, Any]:
+        return {
+            'id': self.id,
+            'online': self.online
+        }
+
+    def deserialize(self, **kwargs) -> None:
+        for k, v in kwargs.items():
+            setattr(self, k, v)
